// my-app/src/app/services/permiso/permiso-interface.ts
export interface PermisoInterface {
    id?: number;
    office_id?: number;
    office_name?: string;
    acronym?: string;
    code?: string;
    mail?: string;
    user_id?: number;
    user_name?: string;
    active?: string;
    worker_id?: number;
    level_id?: number;
    level_name?: string;
}